move address verification class to tcp-server module and subclass in membership
