+++ geode-tcp-server/build.gradle
@@ -25,6 +25,10 @@ dependencies {
 
     implementation('org.apache.logging.log4j:log4j-api')
 
+    //Commons validator is used to validate inet addresses
+    implementation('commons-validator:commons-validator')
+
+
     testImplementation('com.tngtech.archunit:archunit-junit4')
 
     integrationTestImplementation('com.tngtech.archunit:archunit-junit4')
