+++ geode-tcp-server/src/test/resources/expected-pom.xml
@@ -61,5 +61,10 @@
       <artifactId>log4j-api</artifactId>
       <scope>runtime</scope>
     </dependency>
+    <dependency>
+      <groupId>commons-validator</groupId>
+      <artifactId>commons-validator</artifactId>
+      <scope>runtime</scope>
+    </dependency>
   </dependencies>
 </project>
